# Eager-load document uris for activity search