--- a/h/activity/query.py
+++ b/h/activity/query.py
@@ -169,7 +169,9 @@ def _execute_search(request, query, page_size):
 def _fetch_annotations(session, ids):
     return (session.query(Annotation)
             .options(subqueryload(Annotation.document)
-                     .subqueryload(Document.meta_titles))
+                     .subqueryload(Document.meta_titles),
+                     subqueryload(Annotation.document)
+                     .subqueryload(Document.document_uris))
             .filter(Annotation.id.in_(ids))
             .order_by(Annotation.updated.desc()))