@nickstenning 2016-09-08
Am I missing something? Your benchmarks show a 10% improvement on one query but a 300% slowdown on the other.