@nickstenning 2016-09-08
Ok, after discussion in Slack this looks fine.